External pointer should not be generated in initialize
An earlier change to generate the external pointer in the initialize method needed to be undone.  In the initialize method the address of an identifier like Lind is that of the argument.  The field will be a copy of the argument.

More cleanup and insertion of calls to .Internal(inspect(.))